requirements.txt
africastalking==1.2.7
aiohttp==3.9.4
aiohttp-retry==2.8.3
aiosignal==1.3.1
alembic==1.14.0
aniso8601==9.0.1
ansible-base==2.10.8
apache-libcloud==3.2.0
apt-clone==0.2.1
apturl==0.5.2
argcomplete==1.8.1
asgiref==3.5.0
astroid==3.3.5
asttokens==2.4.1
async-timeout==4.0.3
attrs==23.2.0
Babel==2.8.0
bcrypt==4.1.2
beautifulsoup4==4.10.0
black==24.10.0
blinker==1.7.0
Brlapi==0.8.3
cachelib==0.13.0
certifi==2020.6.20
cffi==1.16.0
chardet==4.0.0
click==8.1.7
colorama==0.4.4
comm==0.2.1
command-not-found==0.3
configobj==5.0.6
contextlib2==21.6.0
contourpy==1.2.0
cowsay==6.1
crudini==0.9.3
cryptography==3.4.8
cupshelpers==1.0
cycler==0.12.1
dbus-python==1.2.18
debugpy==1.8.1
decorator==5.1.1
defer==1.0.6
dill==0.3.9
distro==1.7.0
Django==3.2.12
dnspython==2.6.1
email_validator==2.1.1
exceptiongroup==1.2.0
executing==2.0.1
eyeD3==0.8.10
filelock==3.6.0
Flask==3.0.2
Flask-Bcrypt==1.0.1
Flask-DB==0.4.1
Flask-Login==0.6.3
Flask-Mail==0.9.1
Flask-Migrate==4.1.0
Flask-MySQLdb==2.0.0
Flask-RESTful==0.3.10
Flask-Session==0.8.0
Flask-SQLAlchemy==3.1.1
Flask-WTF==1.2.1
fonttools==4.50.0
frozenlist==1.4.1
greenlet==3.0.3
grpcio==1.30.2
gunicorn==21.2.0
gyp==0.1
httplib2==0.20.2
idna==3.3
ifaddr==0.1.7
IMDbPY==2021.4.18
importlib-metadata==4.6.4
iniparse==0.4
ipykernel==6.29.3
ipython==8.22.1
isort==5.13.2
itsdangerous==2.1.2
jedi==0.19.1
jeepney==0.7.1
Jinja2==3.1.3
jmespath==0.10.0
jupyter_client==8.6.0
jupyter_core==5.7.1
keyring==23.5.0
kiwisolver==1.4.5
launchpadlib==1.10.16
lazr.restfulclient==0.14.4
lazr.uri==1.0.6
lockfile==0.12.2
louis==3.20.0
macaroonbakery==1.3.1
Mako==1.1.3
MarkupSafe==2.1.5
matplotlib==3.8.3
matplotlib-inline==0.1.6
mccabe==0.7.0
migrate==0.3.8
more-itertools==8.10.0
msgspec==0.19.0
multidict==6.0.5
mypy-extensions==1.0.0
mysql-connector-python==8.4.0
mysqlclient==2.2.4
nemo-emblems==5.8.0
nest-asyncio==1.6.0
netaddr==0.8.0
netifaces==0.11.0
ntlm-auth==1.4.0
numpy==1.26.4
oauthlib==3.2.0
onboard==1.4.1
packaging==24.1
PAM==0.4.2
pandas==2.2.2
parso==0.8.3
pathspec==0.12.1
pexpect==4.8.0
Pillow==9.0.1
platformdirs==4.2.0
prompt-toolkit==3.0.43
protobuf==3.12.4
psutil==5.9.0
psycopg2-binary==2.9.9
ptyprocess==0.7.0
pure-eval==0.2.2
pycairo==1.20.1
pycparser==2.22
pycryptodomex==3.11.0
pycups==2.0.1
pycurl==7.44.1
pyelftools==0.27
Pygments==2.17.2
PyGObject==3.42.1
PyICU==2.8.1
pyinotify==0.9.6
PyJWT==2.3.0
pykerberos==1.1.14
pylint==3.3.1
pymacaroons==0.13.0
pymongo==4.8.0
PyMySQL==1.1.0
PyNaCl==1.5.0
pyparsing==2.4.7
pyparted==3.11.7
PyQt5==5.15.10
PyQt5-Qt5==5.15.2
PyQt5-sip==12.13.0
pyRFC3339==1.1
PySide6==6.7.2
PySide6_Addons==6.7.2
PySide6_Essentials==6.7.2
python-apt==2.4.0+ubuntu4
python-dateutil==2.8.2
python-debian==0.1.43+ubuntu1.1
python-decouple==3.8
python-dotenv==1.0.1
python-gnupg==0.4.8
python-magic==0.4.24
python-xlib==0.29
pytz==2022.1
pywinrm==0.3.0
pyxdg==0.27
PyYAML==5.4.1
pyzmq==25.1.2
reportlab==3.6.8
requests==2.25.1
requests-file==1.5.1
requests-kerberos==0.12.0
requests-ntlm==1.1.0
requests-toolbelt==0.9.1
schema==0.7.5
scipy==1.13.0
SecretStorage==3.3.1
selinux==3.3
setproctitle==1.2.2
shiboken6==6.7.2
simplejson==3.17.6
six==1.16.0
soupsieve==2.3.1
SQLAlchemy==2.0.30
SQLAlchemy-Utils==0.41.2
sqlparse==0.4.2
stack-data==0.6.3
systemd-python==234
tinycss2==1.1.1
tldextract==3.1.2
tomli==2.0.2
tomlkit==0.13.2
tornado==6.4
traitlets==5.14.1
twilio==9.0.4
typing_extensions==4.12.2
tzdata==2024.1
ubuntu-drivers-common==0.0.0
ufw==0.36.1
Unidecode==1.3.3
urllib3==1.26.5
wadllib==1.3.6
waitress==3.0.0
wcwidth==0.2.13
webencodings==0.5.1
Werkzeug==3.0.4
wordcloud==1.9.3
WTForms==3.1.2
xdg==5
xkit==0.0.0
xlrd==1.2.0
xmltodict==0.12.0
yarl==1.9.4
youtube-dl==2021.12.17
zipp==1.0.0
alembic==1.14.0
bcrypt==4.2.1
blinker==1.9.0
cachelib==0.13.0
cffi==1.17.1
click==8.1.8
colorama==0.4.6
cryptography==44.0.0
decouple==0.0.7
dnspython==2.7.0
email_validator==2.2.0
Flask==3.1.0
Flask-Bcrypt==1.0.1
Flask-Login==0.6.3
Flask-Mail==0.10.0
Flask-Migrate==4.1.0
Flask-Session==0.8.0
Flask-SQLAlchemy==3.1.1
Flask-WTF==1.2.2
greenlet==3.1.1
idna==3.10
itsdangerous==2.2.0
Jinja2==3.1.5
Mako==1.3.8
MarkupSafe==3.0.2
msgspec==0.19.0
pycparser==2.22
PyMySQL==1.1.1
pyscrypt==1.6.2
python-decouple==3.8
python-dotenv==1.0.1
SQLAlchemy==2.0.37
typing_extensions==4.12.2
Werkzeug==3.1.3
WTForms==3.2.1
